aider: Update broadcastAnnotated method to log application info to shared user defaults.
# Aider chat conversation:

ASSISTANT: To report the opened applications to the app group, we need to modify the `SampleHandler.swift` file. Specifically, we need to update the `broadcastAnnotated(withApplicationInfo:)` method to log the application info to the shared user defaults.

Here is the *SEARCH/REPLACE* block:

YouveBeenFramedBroadcastExtension/SampleHandler.swift
```swift
<<<<<<< SEARCH
    override func broadcastAnnotated(withApplicationInfo applicationInfo: [AnyHashable : Any]) {
        logEvent("Received application info: \(applicationInfo)")
    }
=======
    override func broadcastAnnotated(withApplicationInfo applicationInfo: [AnyHashable : Any]) {
        if let applicationName = applicationInfo["UIApplicationOpenURLOptionUniversalLinksOnly"] as? String {
            logEvent("Application opened: \(applicationName)")
        } else {
            logEvent("Received application info: \(applicationInfo)")
        }
    }
>>>>>>> REPLACE
```

This change will log the name of the opened application if it is included in the application info. If the application name is not included, it will log the entire application info as before. Please note that the key "UIApplicationOpenURLOptionUniversalLinksOnly" is used as an example, you should replace it with the actual key that corresponds to the application name in your application info.